♻️  Refactor Text component
