+++ client/app/(auth)/sign-in.tsx
@@ -3,7 +3,7 @@ import { Alert } from 'react-native';
 import { Link, Stack } from 'expo-router';
 import { supabase } from '@/lib/supabase';
 import { View } from '@/components/ui/View';
-import { Text } from '@/components/ui/Text';
+import { Text } from '@/components/ui/Text/Text';
 import { TextInput } from '@/components/ui/Input/Input';
 import { Button } from '@/components/ui/Button/Button';
