+++ client/app/(user)/ui-demo.tsx
@@ -4,7 +4,7 @@ import React, { useState } from 'react';
 import { ScrollView } from 'react-native';
 import { useTheme } from '@/providers/themeProvider';
 import { View } from '@/components/ui/View';
-import { Text } from '@/components/ui/Text';
+import { Text } from '@/components/ui/Text/Text';
 import { Button } from '@/components/ui/Button/Button';
 import { TextInput } from '@/components/ui/Input/Input';
 import { Card } from '@/components/ui/Card/Card';
