+++ client/components/ui/Input/Input.tsx
@@ -8,7 +8,7 @@ import {
 } from 'react-native';
 import { useTheme } from '@/providers/themeProvider';
 import { View } from '../View';
-import { Text } from '../Text';
+import { Text } from '../Text/Text';
 import { ThemedInputProps } from './Input.interface';
 import { createInputStyles } from './Input.styles';
 import { getInputColors } from './Input.helpers';
